MODPATBLK-13 Reorganize tests
