+++ src/test/java/org/folio/rest/handlers/ItemDeclaredLostEventHandlerTest.java
@@ -1,5 +1,7 @@
 package org.folio.rest.handlers;
 
+import static org.folio.repository.UserSummaryRepository.USER_SUMMARY_TABLE_NAME;
+
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
@@ -8,6 +10,8 @@ import org.folio.domain.OpenLoan;
 import org.folio.domain.UserSummary;
 import org.folio.repository.UserSummaryRepository;
 import org.folio.rest.jaxrs.model.ItemDeclaredLostEvent;
+import org.folio.rest.utils.UserSummaryUtils;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -15,11 +19,17 @@ import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 
 @RunWith(VertxUnitRunner.class)
-public class ItemDeclaredLostEventHandlerTest extends AbstractEventHandlerTest {
+public class ItemDeclaredLostEventHandlerTest extends UserSummaryUtils {
   private static final ItemDeclaredLostEventHandler eventHandler =
     new ItemDeclaredLostEventHandler(postgresClient);
   protected static final UserSummaryRepository userSummaryRepository =
     new UserSummaryRepository(postgresClient);
+
+  @Before
+  public void beforeEach(TestContext context) {
+    super.resetMocks();
+    deleteAllFromTable(USER_SUMMARY_TABLE_NAME);
+  }
 
   @Test
   public void userSummaryShouldBeCreatedWhenDoesntExist(TestContext context) {
